+++ modules/Quarklines_one_t.cpp
@@ -254,8 +254,8 @@ void QuarkLine_one_t<QuarkLineType::Q1>::build_Q1_one_t(const Perambulator& pera
     for(const auto& rnd_id : ric_lookup[qll.id_ric_lookup].rnd_vec_ids){
       const size_t rid1 = rnd_id.first - offset; 
       const size_t gamma_id = qll.gamma[0]; // TODO: hard coded! VERY BAD!!!
-      for(size_t row = 0; row < 4; row++){
-      for(size_t col = 0; col < 4; col++){
+      for(int row = 0; row < 4; row++){
+      for(int col = 0; col < 4; col++){
         Ql[pos][qll.id][rnd_counter].block(row*dilE, col*dilE, dilE, dilE)=
           gamma[gamma_id].value[row] *  
           meson_operator.return_rvdaggerv(qll.id_rvdaggerv, t1, rid1).
@@ -309,8 +309,8 @@ void QuarkLine_one_t<QuarkLineType::Q2V>::build(const Perambulator& peram,
       Eigen::MatrixXcd M = Eigen::MatrixXcd::Zero(4 * dilE, 4 * nev);
       for(const auto& rnd_id : ric_lookup[qll.id_ric_lookup].rnd_vec_ids){
         if(check != rnd_id.first){ // this avoids recomputation
-          for(size_t row = 0; row < 4; row++){
-          for(size_t col = 0; col < 4; col++){
+          for(int row = 0; row < 4; row++){
+          for(int col = 0; col < 4; col++){
             if(!qll.need_vdaggerv_dag)
               M.block(col*dilE, row*nev, dilE, nev) =
                 peram[rnd_id.first].block((t1*4 + row)*nev, (t2*4 + col)*dilE, 
@@ -333,8 +333,8 @@ void QuarkLine_one_t<QuarkLineType::Q2V>::build(const Perambulator& peram,
         for(size_t block_dil = 0; block_dil < 4; block_dil++) {
           const cmplx value = gamma[gamma_id].value[block_dil];
           const size_t gamma_index = gamma[gamma_id].row[block_dil];
-          for(size_t row = 0; row < 4; row++){
-          for(size_t col = 0; col < 4; col++){
+          for(int row = 0; row < 4; row++){
+          for(int col = 0; col < 4; col++){
 
           Ql[pos][qll.id][rnd_counter].
                         block(row*dilE, col*dilE, dilE, dilE) +=
@@ -360,8 +360,8 @@ void QuarkLine_one_t<QuarkLineType::Q2V>::build(const Perambulator& peram,
       Eigen::MatrixXcd M = Eigen::MatrixXcd::Zero(4 * dilE, 4 * nev);
       for(const auto& rnd_id : ric_lookup[qll.id_ric_lookup].rnd_vec_ids){
         if(check != rnd_id.first){ // this avoids recomputation
-          for(size_t row = 0; row < 4; row++){
-          for(size_t col = 0; col < 4; col++){
+          for(int row = 0; row < 4; row++){
+          for(int col = 0; col < 4; col++){
             if(!qll.need_vdaggerv_dag)
               M.block(col*dilE, row*nev, dilE, nev) =
                 peram[rnd_id.first].block((t1*4 + row)*nev, (t2*4 + col)*dilE, 
@@ -384,8 +384,8 @@ void QuarkLine_one_t<QuarkLineType::Q2V>::build(const Perambulator& peram,
         for(size_t block_dil = 0; block_dil < 4; block_dil++) {
           const cmplx value = gamma[gamma_id].value[block_dil];
           const size_t gamma_index = gamma[gamma_id].row[block_dil];
-          for(size_t row = 0; row < 4; row++){
-          for(size_t col = 0; col < 4; col++){
+          for(int row = 0; row < 4; row++){
+          for(int col = 0; col < 4; col++){
 
           Ql[pos][qll.id][rnd_counter].
                         block(row*dilE, col*dilE, dilE, dilE) +=
@@ -423,8 +423,8 @@ void QuarkLine_one_t<QuarkLineType::Q2L>::build(const Perambulator& peram,
       int check = -1;
       for(const auto& rnd_id : ric_lookup[qll.id_ric_lookup].rnd_vec_ids){
         if(check != rnd_id.first){ // this avoids recomputation
-          for(size_t row = 0; row < 4; row++){
-          for(size_t col = 0; col < 4; col++){
+          for(int row = 0; row < 4; row++){
+          for(int col = 0; col < 4; col++){
             if(!qll.need_vdaggerv_dag)
               M.block(col*dilE, row*nev, dilE, nev) =
                 peram[rnd_id.first].block((t1*4 + row)*nev, 
@@ -448,8 +448,8 @@ void QuarkLine_one_t<QuarkLineType::Q2L>::build(const Perambulator& peram,
         for(size_t block_dil = 0; block_dil < 4; block_dil++) {
           const cmplx value = gamma[gamma_id].value[block_dil];
           const size_t gamma_index = gamma[gamma_id].row[block_dil];
-          for(size_t row = 0; row < 4; row++){
-          for(size_t col = 0; col < 4; col++){
+          for(int row = 0; row < 4; row++){
+          for(int col = 0; col < 4; col++){
             Ql[pos][qll.id][rnd_counter].
                         block(row*dilE, col*dilE, dilE, dilE) +=
                value * 
@@ -474,8 +474,8 @@ void QuarkLine_one_t<QuarkLineType::Q2L>::build(const Perambulator& peram,
       int check = -1;
       for(const auto& rnd_id : ric_lookup[qll.id_ric_lookup].rnd_vec_ids){
         if(check != rnd_id.first){ // this avoids recomputation
-          for(size_t row = 0; row < 4; row++){
-          for(size_t col = 0; col < 4; col++){
+          for(int row = 0; row < 4; row++){
+          for(int col = 0; col < 4; col++){
             if(!qll.need_vdaggerv_dag)
               M.block(col*dilE, row*nev, dilE, nev) =
                 peram[rnd_id.first].block((t1*4 + row)*nev, 
@@ -499,8 +499,8 @@ void QuarkLine_one_t<QuarkLineType::Q2L>::build(const Perambulator& peram,
         for(size_t block_dil = 0; block_dil < 4; block_dil++) {
           const cmplx value = gamma[gamma_id].value[block_dil];
           const size_t gamma_index = gamma[gamma_id].row[block_dil];
-          for(size_t row = 0; row < 4; row++){
-          for(size_t col = 0; col < 4; col++){
+          for(int row = 0; row < 4; row++){
+          for(int col = 0; col < 4; col++){
             Ql[pos][qll.id][rnd_counter].
                         block(row*dilE, col*dilE, dilE, dilE) +=
                value *
